:building_construction: moved some methods to parent class
